+++ src/Symfony/Component/DependencyInjection/Compiler/ResolvePrivatesPass.php
@@ -11,10 +11,14 @@
 
 namespace Symfony\Component\DependencyInjection\Compiler;
 
+trigger_deprecation('symfony/dependency-injection', '5.2', 'The "%s" class is deprecated.', ResolvePrivatesPass::class);
+
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
 /**
  * @author Nicolas Grekas <p@tchwork.com>
+ *
+ * @deprecated since Symfony 5.2
  */
 class ResolvePrivatesPass implements CompilerPassInterface
 {